Using `::` instead of `:` in type signatures
Haskell, unlike most other ML-style languages, uses `::` for type
signatures and `:` for list cons. This allowed the following sort of
layout:

```hs
function
  :: Foo -- ^ docstring for Foo argument
  -> Bar
  -> Baz
```

But the advent of `-XLinearTypes` in GHC 9.x means that arrows can be
more than two characters long. Also, formatters like `ormolu` generate
the following style, which is more greppable and increasingly popular:

```hs
function ::
  -- | Docstring for Foo argument
  Foo ->
  Bar ->
  Baz
```

So: it's a wart that's different from the rest of the language family,
the alignment style it enables is falling out of favour, and if you
use linear types the alignment dream is dead anyway. Sadly, I think we
have to call it a mistake.